remove per player shot lists. add per player "Charge Hole" type shot slots that track reloads. allow for per player shot slot limits. add a global shot management system and global shot IDs. have the server keep track of shots in a shot manager. remove GM updates from the client. allow the server to update any shot. allow unlimited world weapons of any team color. allow unlimited player shots in the world, independent of shot slots. use the server state to verify all shot parameters. remove redundant info from kill messages. unsalted shots are better for you. add many bugs. write a small novel for a commit message.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/branches/v2_99_shot_branch@16947 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ include/ShotUpdate.h
@@ -46,7 +46,7 @@ struct ShotUpdate {
 
   public:
     PlayerId		player;			// who's shot
-    uint16_t		id;			// shot id unique to player
+    int			id;			// global Shot ID on outbound shots, local shotID on inbound
     float		pos[3];			// shot position
     float		vel[3];			// shot velocity
     float		dt;			// time shot has existed
